+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CMT
 
+## 0.4.1
+
+– Added function for sampling images conditioned on component labels.
+– Extended MCGSM by additional parameters.
+
 ## 0.4.0
 
 - Added spike-triggered mixture model (STM).
@@ -7,6 +12,7 @@
 - Added generalized linear model (GLM) and fully-visible belief network (FVBN).
 - Added mixture of Gaussian scale mixture (MoGSM).
 - Added *PatchMCGSM*.
+- Extended MCGSM by linear features and means.
 - Made implementation of new conditional models easier by introducing interface *Trainable*.
 - Most methods of *MCGSM* can now cope with zero-dimensional inputs.